requirements.txt
Flask==2.0.1
opencv-python==4.5.3.56
mediapipe==0.8.7
numpy==1.21.0
azure-storage-blob==12.8.1
